Handles removal of Bazel packages with open config windows.
- Warns the user if they try to remove Bazel packages while config
  editors are opened, closing the windows if the user decides to
  proceed.

This is "necessary" as config editors may be reliant on Bazel rules
that are contained by the packages being removed and it doesn't seem
worthwhile to try to back all associated config wizards to the first
page.

When config files are loaded, they already print warning messages if
they rely on rules that do not exist in the project.

--
MOS_MIGRATED_REVID=115062857