Disables Swift optimization for release builds.
When Tulsi is built in release mode with Xcode 7.2 it appears to crash
occasionally during project generation. This may have something to do
with the use of generics in PBXProjSerializer and similar crashes have
been found in Swift 2.1 builds in the past.

For now the simplest workaround is to disable optimization until Xcode
7.2 is no longer supported.

--
MOS_MIGRATED_REVID=121152236