Simplifies supporting file extraction.
A large number of Bazel attributes identify files that are nice to
include in the generated Xcode project but do not require any sort of
special handling in the pbxproj itself (xibs and storyboards are good
examples). Rather than create N-different struct members, these
attributes are now resolved and returned in a single list under
"supporting_files."

As a bonus, a bunch of additional attributes are now supported.

--
MOS_MIGRATED_REVID=123344452