Fixes mismatched processing task count.
The way that processing task counts are propagated from the project
level to the config level opened up the potential for a race
condition:

-- ProjectTaskCount = 1
--- Open a generator config, ConfigTaskCount == 0
---- Project task finishes, ProjectTaskCount-- and ConfigTaskCount--
which triggers an assert as task counts cannot be < 0.

This change propagates the project level task count at the moment that
config documents are associated with it.

--
MOS_MIGRATED_REVID=123683761