Removes workspaceroot command-line param.
Previous changes removed the ability to reference config files outside
of the context of a tulsiproj (which already contains workspaceroot
info), so this flag is now duplicative and confusing.

Bonus: Removes some of the isMainThread queue avoidance logic which
has recently been pointed out as being unsafe (and triggers errors in
the case of a headless build anyway, where NSThread.isMainThread is
not the thread that dispatch_main uses to run the main queue).

--
MOS_MIGRATED_REVID=124267694