Generates Xcode schemes for test_suite targets.
test_suite rules aggregate multiple ios_test targets. Currently, any
test_suite selected by the user is simply expanded to its set of tests
and is functionally equivalent to selecting each test individually.
This change generates test_suite-specific xcschemes, allowing the
tests to be executed as an aggregate block.

This approach only works for tests with a common host. If a test_suite
contains tests with multiple hosts the behavior is undefined. In the
current implementation, an arbitrary one will be selected and any
tests that do not target that host will be omitted from the scheme,
but it is not safe to depend on this behavior.

--
MOS_MIGRATED_REVID=124735517