Resolves threading issue in headless mode.
Error/Warning/Info messages printed out by the TulsiProjectDocument
were bounced through the main queue, sometimes leading to them not
being displayed before the program exited due to an exceptional
condition. Furthermore, error messages attempted to spawn an alert
popup despite being in a console environment. This change allows the
message logger used by the project document to be overridden.

This fix is not a particularly good one, in the future it'd be better
to refactor things to split the logging behavior out of TPD and allow
it to be initialized elsewhere.

--
MOS_MIGRATED_REVID=125696361