Adds mixed state handling for recursive checkbox.
Previously there was no indication if a parent node contained a child
that had recursive inclusion enabled. This allows mixedState on
recursive checkboxes such that top level nodes will indicate whether
any of their descendants are recursive-enabled.

Note that it is possible to enable recursive explicitly on all
children of a node and have that node not be recursive itself - this
change does not collapse recursive settings and the behavior is
intentional. Users may wish to recursively include the contents of
some set of children that exist today without automatically including
siblings added in the future.

--
MOS_MIGRATED_REVID=125776587