Adds /* suffix to bundles in generator manifest.
Previously bundles were treated as a mix of (partial) references to
bundle contents (xcdatamodeld bundles had references to the inner
xcdatamodel items but not the xccurrentversion metadata) or to
the bundle itself without any indication that it is a directory
structure. This change appends "/*" to bundle references so users of
the manifest know that the full contents of the bundle are needed.

--
MOS_MIGRATED_REVID=126813062