Fixes break due to 'binary' attribute change.
Recent versions of Bazel have started returning a list of targets
rather than a single target for the ios_application's 'binary'
attribute. It's not immediately clear if this is an intentional change
or not, but this change supports both the unexpected list and the
expected single-value cases.

--
MOS_MIGRATED_REVID=128372559