Reenables --workspaceroot command-line param.
This change restores the "workspaceroot" parameter, though it is now
optional rather than required. This allows the value stored in the
Tulsi project file to be overridden, which may be useful if a user
wishes to share a single Tulsi project for multiple copies of the
same source code (e.g., for multiple local Perforce clients).

--
MOS_MIGRATED_REVID=129439402