Fixes ordering of multiple IPA artifacts.
In certain cases (notably XCTest targets), Bazel may list multiple IPA
files as outputs of a target. This change guarantees that the primary
IPA for a given rule is always first in the list of artifacts provided
to the bazel_build.py script. This allows the script to extract the
primary artifact rather than some side-effect.

--
MOS_MIGRATED_REVID=129654863