Hardcodes include path for STL.
Bazel injects an include path for the bundled version of STL when
compiling C++ code (see CppHelper.java). This injected path does not
seem to be exposed to Skylark at the moment, so this CL hardcodes it
into Tulsi until it can be extracted properly. This requires that the
tools directory be accessible from the workspace root, so symlinking
may be required.
This resolves some odd Live issues errors where seemingly arbitrary
things are flagged as broken because the preprocessor fails on some
(but not all) STL includes.

--
MOS_MIGRATED_REVID=129921794