Injects a guessed IPA name for default test hosts.
XCTest hosts are not required to be included in a project's Tulsi
config. Omitting them, however, leads to an automatically generated
rule entry whose BAZEL_OUTPUTS are not set. At build time, the
bazel_build script is unable to locate the resultant IPA and the
build is failed entirely.

This change guesses the IPA output for automatically generated rule
entries, allowing BAZEL_OUTPUTS to be set appropriately so builds
work properly.

--
MOS_MIGRATED_REVID=129992157