Fixes self-depending merged indexers.
Indexers that depend on each other could be merged into a single
indexer target. When the dependency is resolved, it is translated
into a self-reference, which triggers an assert in debug builds (but
was otherwise correctly ignored).
This change ignores self-references prior to attempting to register
a dependency, fixing the assertion.

Note that this does not guarantee an acyclic graph, but does prevent
self-references. Specifically it is still possible for A and C to be
merged in an A->B->C graph, resulting in a cycle between A_C<->B.

--
MOS_MIGRATED_REVID=130400164