Adds trivial bundle ID extraction.
This change sets PRODUCT_BUNDLE_IDENTIFIER for any rules that set a
bundle ID via the appropriate attributes. A followup change will be
necessary to extract the same setting from Info.plist files in cases
where the ID is not provided by the attribute (which may also require
the Info.plist's to be generated by the build system).

--
MOS_MIGRATED_REVID=132067711