Disables build parallelization for watch2 targets.
I was seeing occasional build failures when building watchOS 2 schemes
that may be attributable to parallel Bazel invocations. Parallelizing
these builds should only save a fraction of a second anyway, as Bazel
will not build multiple targets concurrently in the same workspace, so
the parallelization is disabled as a precaution.

--
MOS_MIGRATED_REVID=132453344