Fixes default-bazel invocation in headless mode.
This change disables eager loading of project documents in commandline
mode whenever the user overrides the default bazel path. This is
necessary as project loading occurs before the bazel path is overridden
on the document, triggering errors if a default bazel path is set but
invalid.

This change also allows paths to be specified relative to home (~)

--
MOS_MIGRATED_REVID=133956372