Global cleanup change.
Also includes the following changes:

Adds intermediateArtifacts to generatorManifest.

This change adds the intermediate dependencies generated as part of
the Bazel build process to the Tulsi generatorManifest file. This is
primarily useful for scripts that need to access these dependencies
(which may contain DWARF symbols in need of patching).

Note: Intermediate dependencies are not currently generated in cases
where dSYM files will be produced (since there is no explicit usecase
and walking the dependency graph to obtain them slows project
generation and produces a significantly larger manifest).

--
Sets keep_going flag on test_suite extractor.

The test_suite Bazel query was unnecessarily strict and could require
the user to have files available that are not actually necessary to
build. This change sets the --keep_going flag to allow some types of
errors to be ignored by the query.

--
MOS_MIGRATED_REVID=137512488