Removes BAZEL_INTERMEDIATE_ARTIFACTS build setting.
The BAZEL_INTERMEDIATE_ARTIFACTS setting was added in order to make it
easy to add intermediate build artifacts to the generated manifest as
well as to allow a future version of post_processor to traverse the
full set of artifacts in order to patch DWARF symbols.

It turns out that for larger projects, this variable can exceed the
maximum allowed size, causing Xcode to silently fail to invoke the
build script.

Since it is not currently important that the build script have
knowledge of intermediate artifacts, this CL removes them altogether.
In the future the build script can either parse the manifest, or the
artifacts can be written to a file whose path is passed via a build
setting.

--
MOS_MIGRATED_REVID=138561187