Sets the project-level SDKROOT.
A previous change intentionally removed the project level SDKROOT as
it caused confusing issues with target device selection for certain
project setups (e.g., a project including both iOS and tvOS targets).
A side effect of this change is that "My Mac" is displayed as a
possible target device regardless of the SDK used by the build target.
e.g., an iOS application would allow users to select "My Mac" even
though the build output was not a macOS binary.

This change fixes the common case by examining all of the build
targets and setting a default SDK if there is an obvious match, only
leaving the SDK unset in cases where it is impossible to express the
default in a reasonable manner (e.g., the tvOS + iOS case above).

--
MOS_MIGRATED_REVID=139474555