Fixes issues with disambiguated targets.
The build script utilizes the PRODUCT_NAME field from Xcode in order
to determine the paths to extract within generated IPAs.
Unfortunately, this field is not entirely safe to use, as Tulsi may be
forced to dramatically change it in order to disambiguate Bazel build
targets with the same name.
This behavior, coupled with an optimization that skips attempts to
install files from the archive whose paths are shorter than the
expected subdirectory + PRODUCT_NAME, may result in a failure to
copy generated artifacts into the CODESIGNING_FOLDER_PATH, triggering
a build failure.

This change removes the use of PRODUCT_NAME.

--
MOS_MIGRATED_REVID=139823266