Wipes out search paths on test runner configs.
Very complex projects can generate HEADER_SEARCH_PATHS that exceed the
limit on environment size, causing Xcode to fail to attempt to build
them.

Luckily, the search path information is entirely unnecessary for the
purposes of building and running the tests, so this CL clears the
setting for the configurations that are used when running tests.

--
MOS_MIGRATED_REVID=139905290