Uses interim aspect flag syntax.
A recent change in Bazel caused the "accidentally working" method used
by Tulsi to run the aspect to go away. In order to support users of
older and newer Bazel versions, the syntax is modified to use the
correct path-based format.

A followup CL will be needed to switch to the new label-based format.

--
PiperOrigin-RevId: 140727243
MOS_MIGRATED_REVID=140727243