Removes the OS_ACTIVITY_MODE=disable default.
OS_ACTIVITY_MODE defaults to "disable" in order to cut down on log
spam in Xcode 8.0. It also unintentionally obliterates device logs,
and the spam issue is theoretically resolved in newer Xcode builds.

This CL reverts the default behavior. Users who need it can opt in by
setting it as an environment flag as normal.

--
PiperOrigin-RevId: 140733180
MOS_MIGRATED_REVID=140733180