Fixes support for watchOS1 apps.
I'm not sure if the requirements of Xcode changed since 6.4 or if
watchOS1 targets were never actually debuggable, but Tulsi was
generating targets that were treated by Xcode as iOS extensions
rather than watchOS1 apps.

This change fixes that, following more or less the same behavior as
that needed for watchOS2 apps (specifically, a customized stub
Info.plist and a do-nothing watch extension).

--
PiperOrigin-RevId: 140865827
MOS_MIGRATED_REVID=140865827