Uses Bazel workspace symlink to resolve paths.
In certain cases Bazel may be able to reference files that are not
direct children of the workspace root. When doing so, it adds symlinks
to those files within the bazel-<workspace_name>/ symlink generated on
build.

This CL utilizes the bazel-<workspace_name>/ symlink in path
references to allow such external resources to be accessed by Xcode.

--
PiperOrigin-RevId: 141070417
MOS_MIGRATED_REVID=141070417