Rollback of commit c8a335456684f46358be6555ed83e6ca7a0f9191.
*** Reason for rollback ***

Rolling back the fix because the root cause of the issue was fixed in []

*** Original change description ***

Remove experimental_enable_objc_cc_deps for
non-ios platforms

Only add the flag when the platform is targeted at iOS, 
since the flag only applies for ios_multi_cpus. The flag currently
does not work when the SDK is set to anything other than ios.

--
PiperOrigin-RevId: 145303432
MOS_MIGRATED_REVID=145303432