Allow use of relative paths for Bazel in scripts
This patch allows the use of a relative path for Bazel in Xcode scripts.

Previously, Tulsi would use an absolute path for Bazel in these scripts when
Bazel was specified relative to the workspace.

This changes the scripts to use the relative path if it is not empty, and default to the full path otherwise.

Example use:

When generating Tulsi projects from the command line, specify a relative path
for Bazel:

$TULSI_APP/Contents/MacOS/Tulsi -- --bazel workspace-relative/path/to/bazel --no-open-xcode --genconfig workspace-relative/to/tulsiproject

All embedded scripts will use the relative path to Bazel as specified by the
command line.