Use additionalSourcePaths if provided on command line
It appears that the `--additionalSourceFilters` command line argument 
is currently ignored. This change takes the simplest approach to 
provision these provided source filters. It assumes that any filter 
provided is intended to be recursive.

A more flexible solution may be desired - I'm open to feedback. In 
the current structure of the code, that would require either parsing 
the arguments (at least to detect if they end in `"/..."`) or 
allowing raw strings to be used instead of or in addition to 
`UISourcePath`s in the document object.

--
Reviewed-on: https://github.com/bazelbuild/tulsi/pull/5
PiperOrigin-RevId: 151735834
MOS_MIGRATED_REVID=151735834