Fix the output matching logic
It was possible for the outputs aspect to match multiple .ipa 
artifacts for ios_test targets where the test target name happened to 
be the suffix of the test rig target. E.g. the rig target 
"app_fooTest" would match when a test target "fooTest" was built.

This change adds a path separator as a boundary so that we
only match the full file name, not part of a path. 

In the aforementioned example, only "/fooTest.ipa" would match.

--
PiperOrigin-RevId: 153611872
MOS_MIGRATED_REVID=153611872