Support dSYM debugging properly
Debugging inside a tulsi generated project currently does not work (see
https://github.com/bazelbuild/tulsi/issues/16 ).

The lldbinit path has always been flaky for me, but it seems like the
dSYM one works well.

The post_processor was not correctly patching the correct paths in the
latest version of tulsi. I imagine that this is due to a change in the
way Bazel builds occur such that there are much more ephemeral paths
that need to be patched.

I attempted to just invoke the post_processor per path, but in large
projects, patching the dSYM was taking close to a minute. This is
unacceptable since the patching needs to run every build.

To fix this, I changed the post_processor to consume a map of path
replacements and then rather than traverse the dSYM per replacement, we
just traverse the dSYM once and check the map at every step. dSYM
replacement now is under 50ms (in fact it takes longer to find the
paths we need to replace).

The logic to find the paths, is located inside a new script
`patchDsym.sh`. This is a separate script because it needs to invoke
`xcrun dwarfdump` and pass the output through many pipes to extract the
proper input for the post_processor map.

```
  Usage: ./patchDsym.sh <dwarfpatcher> \
                        <dsym> \
                        <real-workspace> \
                        <bazel-workspace>
```

The patchDsym.sh takes a path to the dwarfpatcher (the post_processor
binary in this case), the dsym to patch, the path to the project
workspace where a developer does her programming, and the path to the to
root of the bazel workspace.

All other changes support the addition of this script and the new
behavior of the post_processor.

See this branch of a simple bazel project to see debugging working
properly (for normal and external deps):
https://github.com/bkase/debugging-tulsi-broken/tree/working-patcher