Remove usage of Bazel Workspace in XcodeProj
We need to get the Bazel workspace in order to include external sources in the
project. Previously, this was hardcoded to the absolute path of this directory.

In Bazel, this directory is dynamic based on the basename of the workspace root:
Bazel names this directory `bazel-__WORKSPACE_BASENAME__`

Upon building in bazel_build.py, this directory is symlinked to
`tulsi-workspace`. This gives the project generator a stable target.

This replaces all uses of the Bazel workspace with Tulsi workspace and the last
remaining absolute path in the generated project that I could find.

I think with some unit tests and combined with
https://github.com/pinterest/tulsi/pull/2 it could address
https://github.com/bazelbuild/tulsi/issues/18 end to end: the project seems to
be portable based on my testing.

Posting this up as a proposal and request for the Tulsi team's feedback.

Change-Id: I8a689bb426290740bbf439bfa8d6b6ab98b1b718