Remove usage of Bazel Workspace in Xcode Proj
This patch removes the last remaining absolute path in the generated project
that I could find.

We need to get the Bazel workspace in order to include external sources in the
project.

Previously, this was hardcoded to the abosolute path of this directory.

In Bazel, this directory is dynamic based on the basename of the workspace root:
Bazel names this directory `bazel-__WORKSPACE_BASENAME__`

Upon building in bazel_build.py, this directory is symlinked to
`tulsi-workspace`. This gives the project generator a stable target.

Replace all found uses of the Bazel workspace with Tulsi workspace.

Change-Id: I8a689bb426290740bbf439bfa8d6b6ab98b1b718