Use ObjCProvider values for includes and defines
This is better as we let bazel calculate the final values as
opposed to manually calculating the final values and how they
are propagated up on the dependency chain.

This should improve the user experience regarding generated
headers, as the indexers will only contain the include paths
that bazel will make available, instead of all of the
generated header include paths.

--
PiperOrigin-RevId: 161704850
MOS_MIGRATED_REVID=161704850