[RFC] Templating of the runscript
This proposal allows templating of of the runscript action. I'd like to get
feedback on the overall idea and if it it's good prepare for submission :)

I have 2 use case for this:

1) Customize the Bazel invocation dynamically. I want to run static analyzers
and other tools as `extra_actions` which need to be added as an argument to the
Bazel invocation. These actions are conditionally added when the user hits
`analyze` in Xcode. It isn't possible to dynamically change the Bazel invocation
based on the environment otherwise.

2) Override environment variables in Tulsi. Previously, I was just hardcoding the
generator to do this, which is hacky. See the change which I reverted.

This patch allows the user to specify a template as a TulsiOption, and then, the
Bazel invocation and change directory are subbed out.

Additionally, I think there thing that can't be done exclusively within Bazel that
aren't a good fit for pre/post scripts and need to happen in here - there's
probably more use cases than the 2 I mentioned!

TODO:
- Add tests for this. I haven't figured out how to run the integration tests
  but, perhaps it would be a good fit for this.

Change-Id: Iaa7a9742461022cd05d7aa0eb542c4a0ed3bf4ff