Remove explicit test host target dependency (#12)
Tulsi generates XCUITest schemes that self-thrash when custom flags need
to be added to certain targets.

By including the test host as a target dependency, xcode builds the test
host first (with potentially different flags!) before building the
actual XCUITest through bazel (that still builds the underlying app).

In certain situations, like one of our targets at Pinterest, we want to
add custom flags to our host when this test is used. When the app is an
Xcode target this causes (1) the app to build twice for no reason and
(2) a self-thrash of the bazel cache (aka everytime you build is a clean
build).

I tested this with a bazel-clean and executing the target still works
fine.