Print Process invocations in a copyable way
Any Process invocation with a logger will now have
its command logged as an Info message which is easy
to copy and execute in a shell.

This properly escapes the arguments by surrounding
them with single quotes and properly escapes any
single quotes inside the string.

This means the following will be logged:
- bazel info (previously wasn't logged)
- bazel query
- bazel build

--
PiperOrigin-RevId: 169922594
MOS_MIGRATED_REVID=169922594