Fix Xcode Index Store Resource Starvation Race (#22)
Xcode watches the indexer store for changes. If this directory is populated
during a build, Xcode will starve the build process for usage.

This patch uses 2 index stores: 1 for build time and 1 for after, which Xcode
"knows about". After a build, we sync the Tulsi index store with the Xcode one.

From Xcode's perspective, the files are still put in the same place, just in
a batch rather than incrementally.

Change-Id: I7cdd2476172e6bcee435ebf9a7e33d7e3b09b7e9