Disable index store
It doesn't look like the ASTs are loaded into Xcode. I'm seeing an error in
Xcodes indexer that states the AST was discarded.

It looks like this is causing Xcode to reindex after builds. Additionally, there
is a bug in the script causing builds to fail for integration tests. In this
case, the derived data path seems problematic.

We should make sure this does work end to end before reenabling.

Change-Id: I955d804fcbcedf19a02be95b3e25585c65f8cef6