Use the execroot retrieved at project generation.
This avoids the extra cost of making a bazel info execution_root call
during a build, though it does lose a degree of fidelity. In
particular, we lose some ability to correct the execution root if it
changes after a Bazel update, outside of regenerating the Xcode
project.

To fall back, we still query Bazel for the execroot if it is not
found on the filesystem. As it's created during project generation,
this is not expected to happen frequently, unless the user
deliberately cleaned out their Bazel cache.

--
PiperOrigin-RevId: 177826706
MOS_MIGRATED_REVID=177826706