Set include paths for Xcode test targets
This should improve indexing in Xcode test targets.

Note that having the test sources added to the test targets
themselves is a hack and thus we don't expect this to be working
perfectly. Specifically, we don't respect the defines nor copts
as we'd have to scope them per file.

--
PiperOrigin-RevId: 178011515
MOS_MIGRATED_REVID=178011515