Improve ObjC module support in Swift
This improves issues seen in Xcode 8 and 9 where a user is unable to CMD+click
into an ObjC Module while in Swift. Most targets should now be able to be
CMD+click'ed into after the first build.

Note that in some cases Xcode may need to be restarted for the modules
to be seen after building.

--
PiperOrigin-RevId: 178397669
MOS_MIGRATED_REVID=178397669