Refactoring lldbinit codepath target source map.
As we've confirmed that this code path can't handle multiple source
mappings, isolating this to the mapping with the execroot.

LLDB's target.source-map does a string replacement with each
breakpoint that matches the first pair in the list of target.
source-maps. This behavior is not suitable to multiple path mappings.

PiperOrigin-RevId: 182267698