Workaround for saving tulsigen configs.
This is based on NSDocument's header file documentation, which states
that this write method is called within writeSafely after making sure
through NSFileCoordinator (itself a limited means of locking) that it
is safe to read or write to a temporary directory for atomic writing
of document data.

That's not really necessary for Configs, to be honest.

PiperOrigin-RevId: 182790128