Stops trying to re-sign test bundles on macOS.
https://github.com/bazelbuild/tulsi/issues/36

Unit tests on macOS do not require having been code signed, and a
'vanilla' project setup doesn't setup code signing in this case.

When an unsigned bundle is encountered running tests on macOS (which
is the default setup) an error occurs extracting the signing details
(because it's not signed).

As an alternative to this proposal, it may be better to check for an
existing signature as a gating condition here instead of the current
logic.

A second alternative would be to revisit our use of `SIGNING_ALLOWED`
vs. `SIGNING_REQUIRED` (which are set to `YES` and `NO`, respectively,
in the case this CL addresses).

This change is a strawman which proposes the "minimal and safest"
change necessary to resolve the exception.

PiperOrigin-RevId: 183740801