Add check for Spotlight in build script.
Stops the build with an error if it detects that Spotlight has been
disabled on the startup disk.

Spotlight is required to find dSYM bundles used for stack traces and
symbolication in Instruments, and remapping sources with the
DBGVersion 2 API.

If the user needs to build with Tulsi without Spotlight enabled, a
user defined environment variable called
TULSI_BUILD_WITHOUT_SPOTLIGHT_AT_MY_OWN_RISK can be set to YES in the
Tulsi generated project.

PiperOrigin-RevId: 184205621