Automated rollback of commit deab940eb7c76a1bb02d912e7014aba27a067541.
*** Reason for rollback ***

Discovered DBGVersion 2 was not using the intended code path, and
furthermore adding this mapping might be causing breakpoints to
fail to resolve when the "dotted path" is present.

This was most noticeable in Xcode 9.3 beta 2, where with this path
and with DBGVersion 2 corrected so that it was mapping to 2 and not
"2", breakpoints were always failing to resolve.

Will consider if another path mapping, perhaps a more traditional one
such as "/normalized/workspace/" or even "/var/root/", works better
than "./".

PiperOrigin-RevId: 186355392