Add workaround for LLDB latency.
Essentially attempting to make plist generation as fast as possible,
through recommended XML plist generation, but also slowing down the
time taken before we launch LLDB such that the dSYM bundles are more
likely to have been loaded beforehand.

On some builds, DebugSymbols.framework will be unable to find the
dSYM bundle, which can be resolved through rebuilding... but that is
a horrible solution that we can improve on.

mdimport alone does not work, even though we have confirmed the
indexing of dSYMs with mdfind on the com_apple_xcode_dsym_uuids
attribute. Might rely on DBGShellCommands instead.

Will continue to investigate how, if mdimport is not sufficient, to
ensure that LLDB is able to find all dSYM bundles in time before the
debug session is started.

PiperOrigin-RevId: 186493053