Automated rollback of commit 6989b830b6406ec84db56043ad886db665a3abd1.
*** Reason for rollback ***

This change isn't necessary.

On macOS, all compiled apps have their debug data spread out across
all generated .o files, rather than linking it directly into the
binary. The binary itself has a small debug map to point to these .o
files.

The .dSYM bundle pulls all the information from those .o files into
a common binary structure, but this is typically assumed to be
optional.

https://stackoverflow.com/a/12827463
https://stackoverflow.com/a/33307778 (to a lesser extent)

Meanwhile, this -g0 is ignored as it follows the '-g' to emit debug
symbols. Clang's response to combining these flags is to read the
'-g' first and ignore later flags related to emitting debug symbols.

Because the dSYM bundle is composed directly from the generated .o
files and the debug map, there is not a fantastic way to defer or
mitigate that cost.

Some investigation could be done of the dsymutil -minimize feature
that has been in macOS for some time, and recently upstreamed in
llvm-dsymutil (https://reviews.llvm.org/D42688), but that is not
yet a priority.

PiperOrigin-RevId: 187017593