Removed covmap code that hasn't worked since Xcode 7.
Following reports that this feature has not worked for Xcode 8, and
we are well aware that post_processor has a chance of corrupting
binaries, this is absolutely not something we should be keeping in
Tulsi.

We could look into leveraging LLVM/Clang tooling to address coverage
map path remapping, which doesn't seem to work in Swift, but should
provide some solutions for (Obj-)C(++) in our projects.

PiperOrigin-RevId: 187666858