Added common debug path normalization support.
Requires a wrapped_clang and CROSSTOOL update before this will be
enabled. Right now this is overridden with a
TULSI_DIRECT_DBG_PREFIX_MAP flag.

Currently debug path normalization requires the following:

1) There is no Swift. Swift debug info cannot be normalized with
the Swift compiler, swiftc. This could be possible in a future
version of Swift.

2) Post processor (for Xcode < 9) isn't being used to patch dSYM.

3) TULSI_DIRECT_DBG_PREFIX_MAP isn't enabled.

Removed the extra path mapping feature, as it is not necessary for
path normalization, and will not work well with lldbinit based
debugging as it was written.

This change also allows Swift debugging support WITHOUT dSYMs, for
faster builds, on Xcode 9 and above.

PiperOrigin-RevId: 190839630