Bootstrap LLDBInit on project generation.
This presents an opportunity for us to update ~/.lldbinit or
~/.lldbinit-Xcode so that it can locate ~/.lldbinit-tulsiproj so that
it can do path remappings when they are needed for builds without
dSYM bundles.

Still, Xcode requires that it loads ~/.lldbinit or ~/.lldbinit-Xcode
when Xcode itself launches, otherwise it will never be able to
relocate ~/.lldbinit-tulsiproj until the next time Xcode is
restarted.

This could be further improved by generating a prompt for command
line and GUI to request that users restart Xcode after project
generation completes, although that will have to be addressed
alongside other changes and improvements to the Tulsi GUI.

PiperOrigin-RevId: 191456351