Remove main_group_path from bazel_build.py
This has been equivalent to the workspace root for some time, via
TULSI_WR defined as an Xcode user-defined environment variable.

Since TULSI_WR typically references an absolute path to the WORKSPACE
that the generated Xcode project was based on, or references the Xcode
variable $SRCROOT (for more info see
https://pewpewthespells.com/blog/buildsettings.html#srcroot) both of
which are the foundation of all source code made visible to Xcode, the
use of this variable seems redundant, and the comment about debug
information is significantly out of date.

In addition, -fdebug-prefix-map is not supported on Xcode 7, which is
where -fdebug-compilation-dir was last available.

Given that we've used -fdebug-prefix-map as a public default for some
time, and we don't support Xcode 7, it makes more sense to me to
remove its reference entirely, along with other conditional code
assuming the use of Xcode < 8.

PiperOrigin-RevId: 192176379