Add support for rules_swift targets.
Requires bazelbuild/rules_apple 0.5.0 or higher.

This updates the Tulsi aspect to support the legacy SwiftInfo provider from
rules_apple and the new SwiftInfo provider from rules_swift. The legacy rule
and its support will eventually be removed.

This change also uses the new-style `apple_common.Objc` provider anywhere
that the Objc provider of a Swift target is retrieved. This is safer,
because the new Swift rules *only* propagate it using the new-style
syntax, but providers propagated the old way should still be found using
this form.

PiperOrigin-RevId: 201018964