Print Bazel test errors to the Travis CI job log.
In the current form which Travis CI is setup, the Bazel test errors are printed to a log file in its /private/var/... folder which is inaccessible after a test failure.

See https://travis-ci.org/bazelbuild/tulsi/builds/397969960#L2063 for an example.

Use of this flag will print all test errors to the console instead, making them visible in Travis CI's job log.

Closes #54.

PiperOrigin-RevId: 204375657