Xcode 10 support
This PR temporarily disables dark mode support for Xcode 10 until the views can be fully revised to support dark mode on Mojave, in a later Tulsi.

This PR also removes the use of a range constructor that turns out to be wholly unnecessary, as the half open range operator already creates a Swift Range. It breaks the build in Xcode 10. That might have been an artifact in going from NSRanges to Swift Ranges in TulsiGenerator.

Closes #53.

PiperOrigin-RevId: 206004931