Support multiple remappings for dSYMs
When compiling Swift, we automatically enable dSYMs for an improved
debugging experience.

Previously for dSYMs, we had only one of the following mappings:

- ./ -> workspace_root (only for non-Swift with DebugPathNormalization)
- bazel_exec_root -> workspace_root (otherwise)

We should instead always do the later and optionally do the former
if the DebugPathNormalization is enabled. This allows Objective-C
compilation to be normalized even if Swift compilation isn't.

Note that for non-dSYMs, we only use one of the two path remappings
as only one of them should apply (depending of what flags were used
to compile the Cxx sources). It appears that trying to use both
remappings may not work as lldb does not intelligently handle
multiple remappings of the form:
BuildRootA => SourceRoot, BuildRootB => SourceRoot inside .lldbinit's
target.source-map although dSYM remappings do support them.
PiperOrigin-RevId: 206626005