Propagate Swift module name for simple test targets
This allows simple test targets (those with only a single `swift_library`
dependency) to support running individual tests in Xcode.

Xcode uses the PRODUCT_MODULE_NAME setting to filter which tests to run
during execution. Since the PRODUCT_MODULE_NAME was previously set to the
name of the test target rather than the name of the library target, the
filter would be invalid and no tests would execute.

PiperOrigin-RevId: 212869062