Always build both armv7k and arm64_32 architectures when targeting a physical watchOS device or building an application for a physical iOS device that also includes a watch app.
When you build a watchOS target, Xcode does not provide a way to definitively determine the architecture of a connected watchOS device during the run script phase like it does for iOS devices (via the ARCHS environment variable) so we can't know which architecture to target when performing our bazel build.

To ensure we always have the proper application architecture that needs to be deployed, always build a fat binary that includes the slices for both archs.

Also removes alternate "Building for Apple Watch Series 4" option that is now unnecessary since we're always building for it.

PiperOrigin-RevId: 264886356