Skip creating indexer targets for targets that are direct dependencies of test targets.
The way the xcode project is generated, all sources in targets that are direct dependencies of test targets are added into the Compile Sources phase of the test itself. So creating a separate indexer target for that file just means it will be indexed twice by Xcode, when it really doesn't need to be.

PiperOrigin-RevId: 265800475