Do not assume that CopyOnWrite arguments reside on the root filesystem.
We can only do a CoW if the source and destination paths are on the same
APFS volume. Checking that the root volume is APFS is not sufficient.

TESTED=Manual testing on a MacBook with an APFS root volume and a FUSE
mount; we cannot easily automate a test since we can't assume the types
and mount points of filesystems on our automated test machines.
PiperOrigin-RevId: 273558974